backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pdfplumber==0.11.5
python-multipart==0.0.20
python-docx==1.1.2
httpx==0.28.1
python-dateutil==2.9.0.post0
pydantic==2.10.6
